# config/swagger.yml
swagger: '2.0'
info:
  version: v1
  title: wecard
  description: A wecard API based on Swagger and built using Node.js
host: wecard.azurewebsites.net
schemes:
  - http
  - https
basePath: /api
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
security:
  - Bearer: []
consumes:
  - application/json
  - text/json
produces:
  - application/json
  - text/json
paths:
...
  /namecard:
    get:
      tags:
        - Namecard
      summary: Get all Namecards by User id
      operationId: namecard_getByUserId
      parameters:
        - name: userId
          type: string
          in: query
          required: false
          description: The user id
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Namecard'
    post:
      tags:
        - Namecard
      summary: Create a Namecard
      operationId: namecard_create
      parameters:
        - $ref: '#/parameters/Namecard'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Namecard'
    put:
      tags:
        - Namecard
      summary: Update the Namecard
      operationId: namecard_update
      parameters:
        - $ref: '#/parameters/Namecard'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Namecard'
...
definitions:
...
  Namecard:
    type: object
    properties:
      id:
        type: string
      cardType:
        type: integer
      userId:
        type: string
      avatarUri:
        type: string
      name:
        type: string
      phone:
        type: string
      employment:
        type: array
        items:
          type: object
          properties:
            company:
              type: string
            occupation:
              type: string
      education:
        type: array
        items:
          type: object
          properties:
            school:
              type: string
            major:
              type: string
            grade:
              type: integer
      residence:
        type: string
      hometown:
        type: string
...
parameters:
...
  Namecard:
    name: body
    in: body
    required: true
    description: The Namecard
    schema:
      $ref: '#/definitions/Namecard'
...